requirements.txt
# Core Dependencies
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data and Clustering
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Configuration
pyyaml>=6.0

# Development and Testing
pytest>=7.4.0
jupyter>=1.0.0

# Optional: For better performance
# tqdm>=4.65.0
